move YoloItem class to separate file and inject COCOnames into it
