+++ Assets/Scripts/YoloItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8952af0d2956943dc8b467fc675245ea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
